requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
